refactor: apply loading toast message after modal closed
